Refactor life
