+++ ClientApp/src/common/functions/GetFamilySize.tsx
@@ -1,0 +1,10 @@
+export default function formatFamily(input: Number) {
+    if (input < 3) {
+      return "< 3";
+    }
+    if (input > 5) {
+      return "> 5";
+    } else {
+      return "3 - 5";
+    }
+  };
