+++ ClientApp/src/common/functions/GetGender.tsx
@@ -1,6 +1,6 @@
 import Gender from "common/enums/Gender";
 
-export function getGender(gender: Gender, age: number) {
+export default function getGender(gender: Gender, age: number) {
   if (age === undefined || age < 18) {
     switch (gender) {
       case Gender.Other:
