+++ ClientApp/src/components/admin/AdminTab.tsx
@@ -1,7 +1,7 @@
 import { Tab } from "@material-ui/core";
 import { TabContext, TabList, TabPanel } from "@material-ui/lab";
 import React, { useState } from "react";
-import ScrollToTop from "../../components/shared/ScrollToTop";
+import ScrollToTop from "components/shared/ScrollToTop";
 import CompletedMacro from "./completedConnections/Macro";
 import OverviewMacro from "./dashboard/OverviewMacro";
 import ConnectionSuggesterMacro from "./connectionSuggester/ConnectionSuggesterMacro";
