--- ClientApp/src/components/admin/common/style.css
@@ -1,5 +1,0 @@
-#admin-menu {
-    position: absolute;
-    left: 0;
-    margin: 0 1em;
-}
