+++ ClientApp/src/components/admin/dashboard/RecipientTable.tsx
@@ -16,13 +16,13 @@ import {
   FiberManualRecord,
 } from "@material-ui/icons";
 import * as React from "react";
-import Gender from "../../../common/enums/Gender";
-import { RecipientType } from "../../../components/shared/Types";
-import useStyles from "./Styles";
+import { RecipientType } from "components/shared/Types";
+import useStyles from "components/admin/Styles";
 import EditIcon from '@material-ui/icons/Edit';
 import { useState } from "react";
-import * as Types from "../connectionSuggester/Types";
-import EditFamily from "../../shared/EditFamily";
+import EditFamily from "components/shared/EditFamily";
+import getGender from "common/functions/GetGender";
+import formatFamily from "common/functions/GetFamilySize"
 
 type Props = {
   data: RecipientType[] | [];
@@ -37,7 +37,7 @@ const DatatableRecipient: React.FC<Props> = ({
 }) => {
   const classes = useStyles();
 
-  const [selectedRecipient, setSelectedRecipient] = useState({} as Types.RecipientType)
+  const [selectedRecipient, setSelectedRecipient] = useState({} as RecipientType)
   const [open, setOpen] = useState(false)
   const [selected, setSelected] = useState(-1);
 
@@ -46,43 +46,9 @@ const DatatableRecipient: React.FC<Props> = ({
   }
 
   const mergeRecipientTypes = (recipient: RecipientType) => {
-    return Object.assign({} as Types.RecipientType, recipient);
+    return Object.assign({} as RecipientType, recipient);
   }
   
-  const formatFamily = (input: Number) => {
-    if (input < 3) {
-      return "< 3";
-    }
-    if (input > 5) {
-      return "> 5";
-    } else {
-      return "3 - 5";
-    }
-  };
-
-  const handleGender = (gender: Gender, age: Number) => {
-    if (age < 18) {
-      switch (gender) {
-        case Gender.Other:
-          return "Ukjent";
-        case Gender.Male:
-          return "Gutt";
-        case Gender.Female:
-          return "Jente";
-        default:
-      }
-    } else {
-      switch (gender) {
-        case Gender.Other:
-          return "Ukjent";
-        case Gender.Male:
-          return "Mann";
-        case Gender.Female:
-          return "Kvinne";
-        default:
-      }
-    }
-  };
 
   return (
     <Container>
@@ -144,7 +110,7 @@ const DatatableRecipient: React.FC<Props> = ({
             <div key={person.rowKey}>
               <AccordionDetails>
                 <Typography className={classes.smallColumn}>
-                  {handleGender(person.gender, person.age)}
+                  {getGender(person.gender, person.age)}
                 </Typography>
                 <Typography className={classes.smallColumn}>
                   {" "}
