+++ ClientApp/src/components/institution/FormPerson.tsx
@@ -11,10 +11,10 @@ import {
 import ClearIcon from "@material-ui/icons/Clear";
 import * as React from "react";
 import { FC, useEffect, useState } from "react";
-import { GENDERS } from "../../common/constants/Genders";
-import Gender from "../../common/enums/Gender";
-import InputValidator from "../shared/InputFields/Validators/InputValidator";
-import { isNotNull, isInt } from "../shared/InputFields/Validators/Validators";
+import { GENDERS } from "common/constants/Genders";
+import Gender from "common/enums/Gender";
+import InputValidator from "components/shared/InputFields/Validators/InputValidator";
+import { isNotNull, isInt } from "components/shared/InputFields/Validators/Validators";
 import IFormPerson from "./IFormPerson";
 import MessageDialog from "./MessageDialog";
