+++ ClientApp/src/components/institution/InstitutionForm.tsx
@@ -10,11 +10,11 @@ import HelpOutlineIcon from '@material-ui/icons/HelpOutline';
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
 import * as React from "react";
 import { useState} from "react";
-import { DESSERTS } from "../../common/constants/Desserts";
-import { DINNERS } from "../../common/constants/Dinners";
-import Gender from "../../common/enums/Gender";
-import ApiService from "../../common/functions/apiServiceClass";
-import InputValidator from "../shared/InputFields/Validators/InputValidator";
+import { DESSERTS } from "common/constants/Desserts";
+import { DINNERS } from "common/constants/Dinners";
+import Gender from "common/enums/Gender";
+import ApiService from "common/functions/apiServiceClass";
+import InputValidator from "components/shared/InputFields/Validators/InputValidator";
 import Tooltip from '@material-ui/core/Tooltip';
 import ConfirmationDialog from './ConfirmationDialog';
 import FamilyDialog from './FamilyDialog';
