+++ ClientApp/src/components/shared/InputFields/SelectForm.tsx
@@ -6,7 +6,7 @@ import {
   Select,
   SelectProps,
 } from "@material-ui/core";
-import React, { FC, useEffect, useState } from "react";
+import { FC, useEffect, useState } from "react";
 import { useMemo } from "react";
 import { isMobile } from "../../../common/functions/IsMobile";
