+++ ClientApp/src/components/shared/InputFields/Validators/InputValidator.tsx
@@ -1,5 +1,4 @@
 ﻿import { TextField } from "@material-ui/core";
-import * as React from "react";
 import { FC, useEffect, useState } from "react";
 import SelectForm from "../SelectForm";
